Fix for bug when parsing attributes without a name
There is a bug when parsing HTML tags with attributes wihout a name --
either because the author forgot to include the attribute's name (it
happens!) or because they forgot the whitespace between the tagname
and the attribute-name:

<body ="black">
or
<liclass="first">

Parsing this would cause an exit with a function-clause error.

This patch fixes it in such a way that at least the parse succeeds,
although the parsed attribute-names and values are probably not what was
intended by the HTML author.