Remove completely unusued application infrastructure.
There are a number of examples in otp of libraries that have .app files w/o a
mod section.  So it appears that it is perfectly valid to not keep around a
bunch of erlang/otp application infrastructure that is never used.

There is also as has been pointed out many times no good reason to ever
call application:start(mochiweb) and the current incarnation of
mochiweb:start() didn't do anything correctly anyway because we have a
dependence on the inets application which we don't start.

The most reasonable thing to do seems like to remove this infrastructure and
only care about the .app file as it's needed by rebar and other tools like
dialyzer.