Add a simple implementation of max connections
max limits the sum of active connections and listeners.
The implementation assumes that max is greater than acceptor_pool_size,
and always maintains the acceptor_pool at acceptor_pool_size as long
as the max is not exceeded.

If max is less than acceptor_pool_size, acceptor_pool_size is the
effective max.