Support for serving symlinked files.
This commit adds support in mochiweb_request:serve_file/4 for serving files
that are symlinked under DocRoot. The meat and potatoes of it is found in the
function mochiweb_util:interpret_path_links/2, and it should be pretty
self-explanatory. Needless to say, it is a pretty dumb function that happily
reinterprets paths for every request, so I suppose a form of memoization is
needed.